requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0
prophet>=1.1.5
xgboost>=2.0.0
lightgbm>=4.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
schedule>=1.2.0
websockets>=12.0
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0
plyer>=2.1.0
pystray>=0.19.0
Pillow>=10.0.0
tkinter-tooltip>=2.1.0
ttkbootstrap>=1.10.0
bilibili-api-python>=9.1.0
qrcode>=7.4.0
pillow>=10.0.0
